file: Add minimal support for creating new files
We need to be able to initialise the file header and type, as well as
initialise messages to be "all invalid".

I'm not totally sold on the approach of invalidating messages, it would
be better to allocate them and return them to the caller, but that means
implementing getters for every message type on every file type, I think.