codegen: Include length in 'field' struct
For strings and arrays, the size of a field isn't fully defined by its
BaseType. Instead, store a length value in the field array, to be used
by the writer.

For arrays where the size is defined globally (in the "Array" column of
the spreadsheet), use that. Otherwise, if Array == "[N]", or the type is
a string, take the value from the EXAMPLE column, which defines the
array/string length in that particular profile.

The decoder can likely safely ignore this, and just use whatever size is
defined in the file itself.