file_types: Allow for multiple DeviceInfoMsgs
In the SDK specification, it indicates that some files can have multiple
DeviceInfo messages, e.g. to track battery voltage over time, or for
multiple devices.

Add DeviceInfoMsg arrays to the files which allow it, and store
DeviceInfo messages to both the top-level and the body structures.

Note that if there are multiple DeviceInfo messages in a file, the
File.DeviceInfo will be overwritten with the last encountered one
(existing behaviour), whereas the individual file types can store an
array.